Get visibilities from shared objects to use them in filtering
This is the final step in selective auditing: Obtaining the symbol
visibility info, and then deciding if this symbol is an ABI3 violation
with the found visibility.

The rationale is that the non-ABI3-tagged symbols which we know can be
legal are all static inline, which means they should show up as local
in a symbol table.