Add Mach-O bitmask checks for symbol visibility
To the best of my knowledge, taken from derivative works on the Mach-O
documentation found in Google Search.

One nice side effect of these bit hacks is that we can address the TODO
in the loop, because the external marker check is nothing more than
seeing if the first bit in N_TYPE is set.