Get visibilities from shared objects to use them in filtering (#89)
* Get visibilities from shared objects to use them in filtering

This is the final step in selective auditing: Obtaining the symbol
visibility info, and then deciding if this symbol is an ABI3 violation
with the found visibility.

The rationale is that the non-ABI3-tagged symbols which we know can be
legal are all static inline, which means they should show up as local
in a symbol table.

* Add Mach-O bitmask checks for symbol visibility

To the best of my knowledge, taken from derivative works on the Mach-O
documentation found in Google Search.

One nice side effect of these bit hacks is that we can address the TODO
in the loop, because the external marker check is nothing more than
seeing if the first bit in N_TYPE is set.

* Inline dict lookup for ELF visibilities

Also simplify the startswith() check for Python symbols.

* Revert to nameset of allowed symbols, disallow non-ABI3 global symbols

Also, Windows symbol visibilities are all assumed global, because of
the way Windows library exports work.

* Add auditing unit test on cryptography wheels

Downloads wheels from PyPI, caching them for local development, and
audits them across platforms. For one version only, currently v42.0.7.

* Exclude shared object entrypoints from audit reports

These are global, but do not have predetermined names - a shared object's
entry point symbol is always named PyInit_$EXTNAME, with $EXTNAME being
the C++ extension name.

As such, we exclude them in a separate if-branch in the iteration over
all symbols. Since there should only be one entry point per extension,
we expect to find at maximum one match.