cli: Render both mismatches and violations in the same table
Previously, when violations were found (i.e. non-ABI3 symbols in an ABI3-tagged
extension), any subsequent version mismatches (i.e. using symbols that were made
part of the stable ABI, but in later versions than the provided wheel was built for),
were left unreported.

This is now fixed, by changing the audit rendering hook to include all offending
symbols and their ABI versions. In case of a violation, the "not ABI3" value is
given as ABI version, to clearly signal that the found symbol is not part of any
version of the stable ABI.